requirements.txt
ebooklib
google-genai
beautifulsoup4
dearpygui
pywin32
pywinstyles
black